requirements.txt
asgiref==3.5.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.6.0
demjson3==3.0.5
diff-match-patch==20181111
Django==3.2.12
django-cors-headers==3.11.0
django-crispy-forms==1.9.0
django-filter==21.1
django-formtools==2.2
django-guardian==2.2.0
django-import-export==2.0.2
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.17
django-timedeltafield==0.7.10
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
drf-yasg2==1.19.4
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.9
inflection==0.5.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
MarkupPy==1.14
MarkupSafe==1.1.1
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.3
requests==2.23.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
self==2020.12.3
six==1.16.0
sqlparse==0.3.1
tablib==1.1.0
tyadmin-api-cli==0.8.2
uritemplate==3.0.1
urllib3==1.25.8
xadmin-py3==2.0.4
xlrd2==1.3.4
xlsxwriter==3.0.6
xlwt==1.3.0
